Linear Search on Array of Strings in C with example
Your very first data structures practical application with Linear search.
I have explained the Pattern matching algorithm in C in unison with the Linear Search algorithm.
The algorithm of this program can be further divided into sub-algorithms of Linear search and Pattern matching.
Watch the video till the end it might get a bit boring a bit but it is worth the time. I promise that your concepts will be perfectly cleared after watching this video.